requirements.txt
certifi==2018.11.29
chardet==3.0.4
docopt==0.6.2
humanfriendly==4.17
idna==2.8
Pillow==5.4.1
pipreqs==0.4.9
pyreadline==2.1
requests==2.21.0
selenium==3.141.0
textwrap3==0.9.2
urllib3==1.24.1
yarg==0.1.9
